chore: ProfileDetailsForm added
